requirements.txt
python-dotenv
requests
streamlit>=1.31.0
groq
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.10
langchain-core>=0.1.0
langchain-groq
wikipedia
arxiv
tavily-python
